requirements.txt
fastapi
Pillow
numpy
opencv-python
typing 
typing-extensions
uvicorn
python-multipart